refactor(modifynetworkrequest)!: remove non-working Router field from ModifyNetworkRequest
BREAKING CHANGE: The API endpoint allows us to detach the router attached to the network by passing in an empty string ("") or null in the JSON. However, due to the definition here, we can't really pass in either without breaking current consumers of the SDK.

In particular, the problem is that if we do refactor ModifyNetworkRequest to actually pass in "" or null if the field is empty in a fashion that will keep the definition (and not break backwards compatibility), it will cause currently working code that doesn't touch the router to actually detach the routers without any notice to the user. Thus, we need to break backwards compatibility here.